+++ Sources/TuistKit/Commands/Cache/CacheWarmCommand.swift
@@ -16,7 +16,11 @@ struct CacheWarmCommand: ParsableCommand {
     @OptionGroup()
     var options: CacheOptions
 
-    @Argument(help: "A list of targets to cache. Those and their dependent targets will be cached.")
+    @Argument(help: """
+    A list of targets to cache. \
+    Those and their dependant targets will be cached. \
+    If no target is specified, the project defined targets and their dependencies will be  cached.
+    """)
     var targets: [String] = []
 
     @Flag(help: "If passed, the command doesn't cache the targets passed in the `--targets` argument, but only their dependencies")